+++ fltk/src/draw.rs
@@ -614,6 +614,7 @@ pub fn rtl_draw(txt: &str, x: i32, y: i32) {
 /// The string must contain groups of 4 letters which specify one of 24 standard grayscale values, 
 /// where 'A' is black and 'X' is white. The order of each set of 4 characters is: top, left, bottom, right.
 pub fn draw_frame(string: &str, x: i32, y: i32, width: i32, height: i32) {
+    assert!(string.len() % 4 == 0);
     let s = CString::safe_new(string);
     unsafe { Fl_frame(s.as_ptr(), x, y, width, height) }
 }
@@ -621,6 +622,7 @@ pub fn draw_frame(string: &str, x: i32, y: i32, width: i32, height: i32) {
 /// Draws a series of line segments around the given box
 /// Differs from frame() by the order of the line segments which is bottom, right, top, left.
 pub fn draw_frame2(string: &str, x: i32, y: i32, width: i32, height: i32) {
+    assert!(string.len() % 4 == 0);
     let s = CString::safe_new(string);
     unsafe { Fl_frame2(s.as_ptr(), x, y, width, height) }
 }